+++ tasks/unit-tests.js
@@ -1,27 +1,23 @@
 'use strict';
 var gulp        = require('gulp'),
-    babel       = require('gulp-babel'),
     jasmine     = require('gulp-jasmine'),
     runsequence = require('run-sequence');
 
 /**
  * This task runs your unit tests directly in node using jasmine
  * No kharma or phantomJs necessary
- * Compiles sources and test from ES6 to ES5 via babel before running them
+ * Uses webpack to build, witches main entry from application to test index
  */
 module.exports = gulp.task('unit-test', function (cb) {
+    //Set build process to test mode
+    global.config.buildProcess.isTestBuild = true;
 
-    gulp.task('build-source-for-tests', function () {
-        return gulp.src(global.config.paths.source.scripts)
-            .pipe(babel({sourceMaps: false}))
-            .pipe(gulp.dest(global.config.folders.tempTests));
-    });
+    //Change default config to use test sources instead
+    var mainScript = global.config.filenames.main.scripts.valueOf(),
+        tempScript = global.config.filenames.temp.scripts.valueOf();
 
-    gulp.task('build-tests', function () {
-        return gulp.src(global.config.paths.source.tests)
-            .pipe(babel({sourceMaps: false}))
-            .pipe(gulp.dest(global.config.folders.tempTests));
-    });
+    global.config.filenames.main.scripts = global.config.filenames.main.tests;
+    global.config.filenames.temp.scripts = global.config.filenames.main.tests;
 
     gulp.task('run-tests', function () {
         return gulp.src(global.config.paths.temp.tests)
@@ -29,10 +25,22 @@ module.exports = gulp.task('unit-test', function (cb) {
     });
 
     runsequence(
-        'build-source-for-tests',
-        'build-tests',
+        'clean',
+        'build-webpack',
         'run-tests',
-        cb
+        function () {
+            //Reset the isTestBuild flag and script names
+            //after everything was executed
+            //as there may be a develop/release build running afterwards
+            //on the same process
+            global.config.buildProcess.isTestBuild = false;
+            global.config.filenames.main.scripts = mainScript;
+            global.config.filenames.temp.scripts = tempScript;
+            //Force reloading of the webpack-config on next build
+            delete require.cache[require.resolve('../webpack.config.js')];
+
+            cb();
+        }
     );
 
 });